% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_dupes.R
\name{assert_duplicates}
\alias{assert_duplicates}
\title{Assert Duplicates}
\usage{
assert_duplicates(x, remove_na = TRUE)
}
\arguments{
\item{x}{vector of values to test for duplicates}

\item{remove_na}{logical indicating whether to remove \code{NA} values - default to TRUE}
}
\value{
NULL if no duplicates in \code{x}, error message otherwise
}
\description{
Assert Duplicates
}
