% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_present.R
\name{assert_present}
\alias{assert_present}
\title{Assert Present}
\usage{
assert_present(available, required)
}
\arguments{
\item{available}{vector of values to test against vector \code{required}}

\item{required}{vector of required values}
}
\value{
NULL if \code{required} values in \code{available}, error message otherwise
}
\description{
Assert Present
}
