% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_true.R
\name{assert_true}
\alias{assert_true}
\title{Assert TRUE}
\usage{
assert_true(expression, msg = NULL)
}
\arguments{
\item{expression}{expression to evaluate whether \code{TRUE} or \code{FALSE}}

\item{msg}{optional message to display if \code{expression} is \code{FALSE}}
}
\value{
NULL if \code{expression} evaluates to TRUE, error message otherwise
}
\description{
Assert TRUE
}
