% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{clear_results}
\alias{clear_results}
\title{Clear errors and results attributes}
\usage{
clear_results(data)
}
\arguments{
\item{data}{A data frame}
}
\description{
It removes assertr_results and assertr_errors attributes.
}
