% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_id}
\alias{generate_id}
\title{Generates random ID string}
\usage{
generate_id()
}
\description{
This is used to generate id for each assertion error.
}
\details{
For single assertion that checks multiple columns, each error log
is stored as a separate element. We provide the ID to allow
detecting which errors come from the same assertion.
}
