% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_uuid}
\alias{is_uuid}
\title{Returns TRUE if value is a UUID}
\usage{
is_uuid(x)
}
\arguments{
\item{x}{a character vector}
}
\value{
A vector of the same length that is TRUE when the element is
a UUID and FALSE otherwise
}
\description{
This functions returns true if the input is a valid uuid.
}
\examples{
is_uuid("xxxx")
is_uuid("de305d54-75b4-431b-adb2-eb6b9e546014")
}

