% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/row-redux.R
\name{maha_dist}
\alias{maha_dist}
\title{Computes mahalanobis distance for each row of data frame}
\usage{
maha_dist(data, keep.NA = TRUE, robust = FALSE)
}
\arguments{
\item{data}{A data frame}

\item{keep.NA}{Ensure that every row with missing data remains NA in
the output? TRUE by default.}

\item{robust}{Attempt to compute mahalanobis distance based on
        robust covariance matrix? FALSE by default}
}
\value{
A vector of observation-wise mahalanobis distances.
}
\description{
This function will return a vector, with the same length as the number
of rows of the provided data frame, corresponding to the mahalanobis
distances of each row.
}
\details{
This is useful for finding anomalous row-wise observations.

It will convert strings into numerics.
}
\examples{
maha_dist(mtcars)

maha_dist(iris, robust=TRUE)


library(magrittr)            # for piping operator

# using every column from mtcars, compute mahalanobis distance
# for each observation, and ensure that each distance is within 10
# median absolute deviations from the median
mtcars \%>\%
  insist_rows(maha_dist, within_n_mads(10), everything())
  ## anything here will run
}
\seealso{
\code{\link{insist_rows}}
}

