% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{not_in_set}
\alias{not_in_set}
\title{Returns TRUE if value in not in set}
\usage{
not_in_set(..., allow.na = TRUE, inverse = FALSE)
}
\arguments{
\item{...}{objects that make up the set}

\item{allow.na}{A logical indicating whether NAs (including NaNs)
should be permitted (default TRUE)}
}
\value{
A function that takes one value and returns TRUE
        if the value is not in the set defined by the
        arguments supplied by \code{not_in_set} and FALSE
        otherwise
}
\description{
This function returns a predicate function that will take a single
value and return TRUE if the value is not a member of the set of objects
supplied. This doesn't actually check the membership of anything--it
only returns a function that actually does the checking when called
with a value. This is a convenience function meant to return a
predicate function to be used in an \code{\link{assertr}} assertion.
}
\examples{
predicate <- not_in_set(3,4)
predicate(5)

## is equivalent to

not_in_set(3,4)(5)

}
\seealso{
\code{\link{in_set}}
}
