% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{print.assertr_assert_error}
\alias{print.assertr_assert_error}
\title{Printing assertr's assert errors}
\usage{
\method{print}{assertr_assert_error}(error)
}
\arguments{
\item{x}{An assertr_assert_error object}
}
\description{
`print` method for class "assertr_assert_error"
This prints the error message and the entire two-column
`data.frame` holding the indexes and values of the offending
data.
}
\seealso{
\code{\link{summary.assertr_assert_error}}
}
