% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{print.assertr_defect}
\alias{print.assertr_defect}
\title{Printing assertr's defect}
\usage{
\method{print}{assertr_defect}(x, ...)
}
\arguments{
\item{x}{An assertr_defect object}

\item{...}{Further arguments passed to or from other methods}
}
\description{
`print` method for class "assertr_defect"
This prints the defect message along with columns that were checked.
}
