% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{print.assertr_verify_error}
\alias{print.assertr_verify_error}
\title{Printing assertr's verify errors}
\usage{
\method{print}{assertr_verify_error}(x, ...)
}
\arguments{
\item{x}{An assertr_verify_error object.}

\item{...}{Further arguments passed to or from other methods}
}
\description{
`summary` method for class "assertr_verify_error"
}
\seealso{
\code{\link{summary.assertr_verify_error}}
}
