% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{summary.assertr_assert_error}
\alias{summary.assertr_assert_error}
\title{Summarizing assertr's assert errors}
\usage{
\method{summary}{assertr_assert_error}(object, ...)
}
\arguments{
\item{object}{An assertr_assert_error object}

\item{...}{Additional arguments affecting the summary produced}
}
\description{
`summary` method for class "assertr_assert_error"
This prints the error message and the first five
rows of the two-column `data.frame` holding the
indexes and values of the offending data.
}
\seealso{
\code{\link{print.assertr_assert_error}}
}
