% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{v_class}
\alias{v_class}
\title{Returns vector of classes for selected dataset}
\usage{
v_class(...)
}
\arguments{
\item{...}{data.frame or set of vectors dicided by \code{,}}
}
\value{
vector of classes for each variable of data.frame, or for each passed variable
}
\description{
Returns vector of classes for selected dataset
}
