% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\docType{class}
\name{Validator}
\alias{Validator}
\title{Class providing object with methods for simple data validation reports.}
\format{\code{\link{R6Class}} object.}
\usage{
Validator
}
\value{
Object of \code{\link{R6Class}} with methods for simple data validation reports.
}
\description{
Class providing object with methods for simple data validation reports.
}
\examples{
validator <- Validator$new()
}
\keyword{data}
