% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/assertr.R
\docType{package}
\name{assertr}
\alias{assertr}
\alias{assertr-package}
\title{assertr: Assertive programming for R analysis pipeline.}
\description{
The assertr package does this and that. It does
\itemize{
  \item this
  \item that
}
}
\details{
See the assertr vignette or the documentation for more information \cr
> \code{vignette("assertr")}
}
\examples{
example
}

