% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{summary.assertr_verify_error}
\alias{summary.assertr_verify_error}
\title{Summarizing assertr's verify errors}
\usage{
\method{summary}{assertr_verify_error}(object, ...)
}
\arguments{
\item{object}{An assertr_verify_error object}

\item{...}{Additional arguments affecting the summary produced}
}
\description{
`summary` method for class "assertr_verify_error"
}
\seealso{
\code{\link{print.assertr_verify_error}}
}
